[cli] Add support for native server deployments during export/eas update
